+++ CHANGES.rst
@@ -13,6 +13,12 @@ Changelog for Websauna
 
 - Update jQuery to 3.1.1. Old jQuery is kept in the source tree for backwards compatibility.
 
+- Added :py:class:`websauna.system.crud.views.CSVListing` and user CSV export
+
+- :ref:`ws-shell` `request.application_url` comes from :ref:`websauna.site_url` setting and does not default to localhost
+
+- Write the session cookie and Redis session data only if something was stored in the session. This makes it more scalable to serve anonymous pages, as one does not need to reset the cookie on every crawling request.
+
 1.0a4 (2017-01-07)
 ------------------